+++ src/alarm/constants.ts
@@ -1,0 +1,4 @@
+export const ALARM_CONFIG_KEY = 'alarmConfig';
+export const ALARM_RUNTIME_KEY = 'alarmRuntime';
+export const LEGACY_ALARM_STATE_KEY = 'alarm';
+